+++ server/src/controllers/index.ts
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const newSession = require('./newSession');
-const dataSession = require('./dataSession');
+const newSession = require('../services/newSession');
+const dataSession = require('../services/dataSession');
 
 router.use('/newSession', newSession);
 router.use('/dataSession', dataSession);
